+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to the "mme2k-powerapps-helper" extension will be documented in this file.
 
+## 0.4.0
+
+* Improved tooltips for tree nodes.
+* Improved usability, when updating OAuth settings.
+* New setting, which allows to specify CustomConnector authentication settings for environments.
+* Solution import asks now for update OAuth settings for all CustomConnectors of the solution.
+* Command **"Update OAuth Settings"** is now available for a solution to update all CustomConnectors (OAuth) of the solution. This makes a solution import much more comfortable.
+* Command **"Publish Customizations"** call now the Crm action `PublishAllXml`
+
 ## 0.3.5
 
 * Fix: Wrong cached client secret was used, when the client id changed during input of Update OAuth Settings